requirements.txt
Django==2.1.7
django-environ=0.4.5
requests=2.21.0
djangorestframework=3.9.1